CHANGE: Create Handler for checking and managing types
