+++ src/Business.Convert/Business.Convert.csproj
@@ -30,8 +30,9 @@
     <Reference Include="System" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClassDiagramConverter.cs" />
+    <Compile Include="Converter\ClassDiagramConverter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TypeHandler.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Business.Api\Business.Api.csproj">
