+++ src/Business.Convert/Converter/ClassDiagramConverter.cs
@@ -1,9 +1,7 @@
 ﻿using Ninkovic.Stefan.CSharpToPlant.Business.Api;
 using Ninkovic.Stefan.CSharpToPlant.Common.Data;
-using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Linq;
 using System.Reflection;
 
 namespace Ninkovic.Stefan.CSharpToPlant.Business.Convert
@@ -13,6 +11,16 @@ namespace Ninkovic.Stefan.CSharpToPlant.Business.Convert
     /// </summary>
     public class ClassDiagramConverter : IConvert
     {
+        private readonly TypeHandler _typeHandler;
+
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        public ClassDiagramConverter()
+        {
+            _typeHandler = new TypeHandler();
+        }
+
         /// <summary>
         /// Goes through every assembly and creates a new instance of <see cref="ProjectAssembly"/> 
         /// including all types, methods and fields.
@@ -27,7 +35,7 @@ namespace Ninkovic.Stefan.CSharpToPlant.Business.Convert
             {
                 var projectAssembly = new ProjectAssembly(assembly);
 
-                foreach (var type in GetTypes(assembly))
+                foreach (var type in _typeHandler.GetTypes(assembly))
                 {
                     if (type == null) continue;
                     var projectType = new ProjectType(type);
@@ -59,46 +67,6 @@ namespace Ninkovic.Stefan.CSharpToPlant.Business.Convert
             }
 
             return projectAssemblies;
-        }
-
-        /// <summary>
-        /// Gets all types from a assembly
-        /// </summary>
-        /// <param name="assembly">assembly to fetch from</param>
-        /// <returns>Collection of types referenced in the assembly</returns>
-        public ICollection<Type> GetTypes(Assembly assembly)
-        {
-            try
-            {
-                return FilterAnonymousTypes(assembly.GetTypes());
-            }
-            // Catches the exception when a type is referenced which is not known in the asembly
-            catch (ReflectionTypeLoadException e)
-            {
-                // Retrieves only the known types
-                return FilterAnonymousTypes(e.Types);
-            }
-        }
-
-        /// <summary>
-        /// Checks each type in the collection if it exists and also if any special characters are included
-        /// </summary>
-        /// <param name="types">Collection to check</param>
-        /// <returns>Filtered collection</returns>
-        public ICollection<Type> FilterAnonymousTypes(Type[] types)
-        {
-            var correctTypes = new Collection<Type>();
-            foreach (var type in types)
-            {
-                if (type == null) continue;
-                else if (type.Name.Contains('<')) continue;
-                else if (type.Name.Contains('>')) continue;
-                else if (type.Name.Contains('\'')) continue;
-                else if (type.Name.Contains('`')) continue;
-
-                correctTypes.Add(type);
-            }
-            return correctTypes;
         }
     }
 }
